linebot19981130/requirements.txt
Flask==1.1.2
line-bot-sdk==1.19.0
requests==2.25.1
bs4==0.0.1
urllib3==1.26.3